rename component to components
